Do not create session without command
This will prohibit the creation of a session id in case there has not
been entered a command. Else it may mess up the sessions and mix their
commands. This will happen if you create two shells and do not fire a
command in between in the newly created one.