histdb-sync: Flush write-ahead log before committing, pulling and merging
As a performance optimization, sqlite's write-ahead log (wal) was
enabled in 873610f1, which means writes are appended to this .db-wal
file  and only written to the actual .db file after a lot of writes or
when the db connection is closed (which, since a4a60e3b7a, happens only
on zsh exit). However, only the .db file is tracked by git, the wal
being nothing but a temporary cache.

This brings a few issues:

- Recent commands, only in the wal, aren't committed by histdb-sync
- If an update is pulled and the .db file modified, but pending changes
  are still in the wal, sqlite gets very confused and interprets it as a
  corrupt database. Fixing this requires closing the db connection (aka
  the whole shell) and deleting the wal, loosing the last bits of
  history for all sessions.
    - Of course, if you have multiple terminals open, you must do it for
      all of them.

Note this still commit merely shortens significantly the histdb-sync/wal
race condition window - a command typed in another terminal during the
sync could still write to the wal and cause issues. However, working
around that is a bit more involved and probably implies locking the
whole database during the sync, which might bring other issues, and at
that point, the question of whether the wal perf. increase is worth it.