Better sqlite termination
Appears to fix the thing where sqlite inherits a writable copy of the
file descriptor for the sqlite pipe, and so doesn't get EOF and
doesn't get terminated. I think this is because cloexec doesn't stick
if you dup2 an fd, and zsh is dup2ing our fd somewhere after we open
it with cloexec.