Add support for merging unrelated Git histories
When merging unrelated Git histories, the merge driver uses an empty
ancestor database. Thus, the SQL `max` function cannot operate and
returns a `null` value during the merge, effectively discarding *our*
database during the merge. This commit ignores the ancestor database if
it is empty, so both histories are full preserved in this corner case.